make DropKeep use only 3 bytes
This optimization makes both, drop and keep, values use only 12 bits each, restricting their range of valid numbers to just 0-4095. While low, these numbers should suffice for practical use cases. If practical use cases arise that exceed these limitations we might think about new or better encodings.

+++ crates/wasmi/src/engine/func_builder/translator.rs
@@ -354,8 +354,8 @@ impl<'parser> FuncTranslator<'parser> {
         let len_params_locals = self.locals.len_registered() as usize;
         DropKeep::new(
             // Drop all local variables and parameters upon exit.
-            drop_keep.drop() + len_params_locals,
-            drop_keep.keep(),
+            drop_keep.drop() as usize + len_params_locals,
+            drop_keep.keep() as usize,
         )
         .map_err(Into::into)
     }
